perf: lazily load README when embedding publish manifest
